updated types of wsclean arguments
the mypy linter had issues with the types supplied to the arguments of
arrakis.utils.wsclean. Particularly when a None was supplied as a
default but the specified type was resolved (e.g. str, float, int). I
think these should be optional.

Similarly, mypy did not like Union[str, None]. I switched these to
Optional[str], which seems to have resolved some of its issues.