simplified workflow / tweaks to cleanup
Attempting to isolate why imaging in consistently restarted. In doing so
have reduced the amount of tasks that are submitted through to the
graph. There were initially some issues with how the graph was
constructed, but with this version they are largely avoided at the cost
of parrallelism (though without the number of workers to service doesnt
matter too much)

There still seems to be failure to properly clean up files. Almost
there.