initial attempt at pulling in the imager pipeline
This is the first pass at bringing in the main Imager pipeline into the overall Arrakis workflow. The general workflow is now divided up into two parts:
1 - Imaging of the data and cube creation (using wsclean)
2 - Cublet extract, combination, RM-synthesis/clean

Stage 1 requires a different cluster configuration, so is separated into its own sub-flow with its own cluster. When this sub-flow is running, the main flow will be in a blockin state.