fix for blended merging/mongodb tint query
The approach to join the blended dataframe and the cat astropy table
somehow stopped working. Rather than failing, it would match the tables
when reindexing, but fill all columns with nans. I believe somewhere
the column being reindexed become a level in a multi-index, which then
failed to be found. It seems really messed up it took this long to
figure out and not happy with the apparent;y fix

The query for total integration time required some tweaks. Somewhere the
field names had the old 'A' and 'B' character suffix. These needed to be
removed. The query was also looking for the `SELECT` field, which, as I
understand, does not exist.